% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expmeasure.R
\name{expmeasure}
\alias{expmeasure}
\title{Open Expmeasure software}
\usage{
expmeasure()
}
\description{
\code{expmeasure} opens a shiny-based software that helps iGEMers to
analyse their data in part characterization experiments.
}
\details{
This function opens opens a shiny-based software that provides a range of
statistical tools which can be useful for part characterization in iGEM.
To get start with the software, please check the user manual and try to
explore the software using test data \code{sPrcn}.
}
\examples{
##Once you execute the following function
##  It will not stop until you close the window of "Expmeasure" GUI
##expmeasure()
}
\author{
Wu Hao-Ran <haoranwu@zju.edu.cn>
}
