% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expmeasure.R
\name{expmeasure}
\alias{expmeasure}
\title{Open Expmeasure software}
\usage{
expmeasure()
}
\description{
\code{expmeasure} opens a shiny-based software that helps iGEMers to
analyse their data in part characterization experiments.
}
\details{
This function opens opens a shiny-based software that provides a range of
statistical tools which can be useful for part characterization in iGEM.
Data used in this software should be database file formats with each column
specifies a variable and each row for an observation.
}
\examples{
##Once you execute the following function
##  It will not stop until you close the window of "Expmeasure" GUI
\dontrun{
expmeasure()
}
}
\author{
Wu Hao-Ran <haoranwu@zju.edu.cn>
}
